+++ Common/Math/curves.cpp
@@ -37,7 +37,7 @@ float linearOut(int t, int fadeOutLength) {
 float ease(float val) {
 	if (val > 1.0f) return 1.0f;
 	if (val < 0.0f) return 0.0f;
-	return ((-cosf(val * PI)) + 1.0f) * 0.5;
+	return (float)(((-cosf(val * PI)) + 1.0f) * 0.5);
 }
 
 float ease(int t, int fadeLength)
